Add accessibility guidelines
Add a set of specific accessibility guidelines for authoring web-based
interfaces to make them accessible for users using a range of devices
and those who rely on assistive technology such as screen readers.

This document covers recurring issues that have come up in code review
and when improving accessibility of front-end content on the website and
in the client.

This also includes links to content that I found useful while working on
some recent improvements to the website.