Add initial support for converting CommonJS/Node exports to ES exports
 - Add a proper CLI interface using commander

 - Support `-i` and `-e` flags to control whether imports and exports
   are converted respectively.

When migrating an existing project it may be desirable to convert only
the imports at first in case the project has some `require` calls which
cannot be converted automatically. Those calls will break if they expect
the call to return a default export directly rather than an object with a
`default` property, as ES modules do.