Strip "use strict" directives when converting to modules
Modules are always strict, so we can strip these out.

This _should_ be done by finding `ExpressionStatement`s containing only
a "use strict" literal in the AST and removing them. However these
statements do not appear in the AST. I think that is because the source
is being parsed as a module. This commit implements a simple Good
Enough™ hack to remove them using regexes.