Update typing and identifier naming sections in js-guide.md
Update the guide to reflect changes in use of TypeScript and identifier names.

When the guide was last updated we were still authoring code in JS / JSX and adding type annotations using JSDoc comments. Since then we switched to using native TypeScript syntax for all of our main projects. Only tooling which runs in Node, some legacy code (eg. in h) and unit tests now uses JS.

For identifier naming, add an exception for objects that are deserialized from API response bodies or will be serialized as API request bodies.