Update and simplify typechecking-faqs.md
Update typechecking-faqs.md to reflect the fact that we now author production code in TypeScript instead of JSDoc comments.

These obsoletes a number of the questions which were focused on dealing with the less-well documented JSDoc syntax.

The section at the end about `useRef` and `useState` types has become obsolete as a result of the switch to TS, which makes typing easier, and also changes in Preact which made the comment about how `useRef<T|null>(null)` works incorrect. The preferred way to use `useRef`/`useState` now is just to use them in the most obvious way like any other function. Hence no special guidance is needed.